* Added support for 64-bit hashes of values   * Reduces probability of hash collision   * Removed load/store prop opcodes which depend on 32-bit hashes

+++ HulaScript24/hash.h
@@ -2,8 +2,8 @@
 
 #include <cstdint>
 
-static uint32_t constexpr str_hash(char const* input) {
+static uint64_t constexpr str_hash(char const* input) {
     return *input ?
-        static_cast<uint32_t>(*input) + 33 * str_hash(input + 1) :
+        static_cast<uint64_t>(*input) + 33 * str_hash(input + 1) :
         5381;
 }
